Fetch and display Mars photos via REST API
Developed a simple application to fetch Mars photos using a REST API as part of the project scope. API calls were made using the Retrofit library, and the retrieved data was displayed in a RecyclerView. Each photo item was mapped with a model class representing the JSON data received from the API. The application provides an interface for users to view photos taken from Mars.